added option for landing zone
Summary: notebooks to give users the option to copy files from synthea_raw_files volume to a landing volume where files are organized into directories by object (e.g. all patient files live in /landing/patients/*). This is to support other potential use cases for streaming, batch and AutoLoader configuration.

Updated readme to include create_landing_zone widget for configuration, along with the purpose of the landing zone volume and the difference in file structure between landing and synthea_raw_files volumes.

Added widget to synthea-on-databicks.py to add option to users to create a landing zone volume. Updated arguments and parameters accordingly.

Added images folder for images used in readme

Added notebook called 2.0-move-synthea-files-to-land.py. This is the core notebook for copying new files in synthea_raw_files volume over to a landing zone volume where files are organized into directories by object rather than by timestamp.

Updated workflow notebook to added condition for the landing zone and to update the job definition with 2.0-move-synthea-files-to-land.py

Updated workflow notebook to get the notebook current base path to dynamically populate the notebook paths in the workflow definition. This is to give users the flexibility to run these notebooks from anywhere in their workspace (e.g. in their Repo folder) rather than requiring the notebooks to be in /Workspace/Users/ or to manually update the paths in the workflow definitions.